# specs/005-cloud-event-deployment/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Chatbot API - Phase V Extensions
  version: 5.0.0
  description: Advanced task features, recurring tasks, tags, search, and event-driven endpoints
paths:
  /api/{user_id}/tasks:
    get:
      summary: List tasks with filters and sorting
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - name: tag
          in: query
          schema:
            type: integer
          description: Tag ID to filter by
        - name: overdue
          in: query
          schema:
            type: boolean
        - name: due_soon
          in: query
          schema:
            type: boolean
          description: Due within 24 hours
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, title, due_date, priority, completed]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Create task with priority, tags, due date
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/{user_id}/tasks/{task_id}:
    put:
      summary: Update task including priority, tags, due date
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/{user_id}/tasks/search:
    get:
      summary: Full-text search tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query for title and description
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [high, medium, low]
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          nullable: true
        recurring_task_id:
          type: integer
          nullable: true
        is_recurring_instance:
          type: boolean
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        created_at:
          type: string
          format: date-time
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        due_date:
          type: string
          format: date
        due_time:
          type: string
        tag_ids:
          type: array
          items:
            type: integer
        recurring:
          type: boolean
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
        recurrence_interval:
          type: integer
        days_of_week:
          type: array
          items:
            type: integer
        day_of_month:
          type: integer
        end_date:
          type: string
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        completed:
          type: boolean
        due_date:
          type: string
          format: date
        due_time:
          type: string
        tag_ids:
          type: array
          items:
            type: integer
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          description: Hex color (#RRGGBB)
        created_at:
          type: string
          format: date-time
...